Put form parsing logic into seperate util, add space update handler
