+++ src/lib/components/SpaceForm.svelte
@@ -1,8 +1,10 @@
 <script lang="ts">
   import AvailabilitySetter from "$lib/components/AvailabilitySetter.svelte";
-  import { create } from "$lib/api/spaces";
+  import { create, update } from "$lib/api/spaces";
+  import { parseSpaceFormData } from "$lib/utils";
 
   let { formType } = $props();
+  let selectedSpaceType = $state("physical");
   let availability: { date: string; startTime: string; endTime: string }[] =
     $state([]);
   let alwaysAvailable = $state(false);
@@ -12,84 +14,21 @@
   ) {
     availability = newAvailability;
   }
+  function handleSubmit(e: Event) {
+    if (formType === "create") {
+      handleCreateSpace(e);
+    } else if (formType === "edit") {
+      handleUpdateSpace(e);
+    }
+  }
+
   async function handleCreateSpace(e: Event) {
     e.preventDefault();
 
     const form = e.currentTarget as HTMLFormElement;
     const formData = new FormData(form);
 
-    // Extract form values
-    const type = formData.get("space-type") as "physical" | "gps" | "virtual";
-    const name = formData.get("space-name") as string;
-    const capacity = parseInt(formData.get("capacity") as string, 10);
-    const accessibility = formData.get("accessibility-details") as string;
-    const description = formData.get("space-description") as string;
-    const contact = formData.get("contact-details") as string;
-    const message = (formData.get("space-message") as string) || "";
-    const multiBookable = formData.get("multi-bookable") === "true";
-
-    // Parse location based on type
-    let location: PhysicalLocation | GPSLocation | VirtualLocation = {
-      street: "",
-      city: "",
-      state: "",
-      zip: "",
-      country: "",
-    }; // Default to PhysicalLocation to satisfy TypeScript
-
-    if (type === "physical") {
-      location = {
-        street: formData.get("address-street") as string,
-        city: formData.get("address-city") as string,
-        state: formData.get("address-state") as string,
-        zip: formData.get("address-zip") as string,
-        country: formData.get("address-country") as string,
-      };
-    } else if (type === "gps") {
-      location = {
-        lat: (formData.get("address-lat") as string).trim(),
-        lon: (formData.get("address-lon") as string).trim(),
-      };
-    } else if (type === "virtual") {
-      location = formData.get("address-virtual") as string;
-    }
-
-    // Parse link
-    const linkTitle = formData.get("space-link-text") as string;
-    const linkUrl = formData.get("space-link-url") as string;
-    const link: Link | null =
-      linkTitle && linkUrl
-        ? { type: "custom" as const, title: linkTitle, url: linkUrl }
-        : null;
-
-    // Parse availability
-    let parsedAvailability: TimeSpan[] | "always";
-    if (alwaysAvailable) {
-      parsedAvailability = "always";
-    } else {
-      parsedAvailability = availability.map(({ date, startTime, endTime }) => ({
-        start: new Date(`${date}T${startTime}`),
-        end: new Date(`${date}T${endTime}`),
-      }));
-    }
-
-    // Prepare the payload
-    const payload: SpaceFields = {
-      type,
-      name,
-      location,
-      capacity,
-      accessibility,
-      description,
-      contact,
-      link,
-      message,
-      images: [],
-      availability: parsedAvailability,
-      multiBookable,
-    };
-
-    console.log("Final Payload:", payload);
+    const payload = parseSpaceFormData(formData, alwaysAvailable, availability);
 
     try {
       const response = await create(payload);
@@ -99,10 +38,26 @@
     }
   }
 
-  let selectedSpaceType = $state("physical");
+  async function handleUpdateSpace(e: Event) {
+    e.preventDefault();
+
+    const form = e.currentTarget as HTMLFormElement;
+    const formData = new FormData(form);
+
+    const spaceId = "1"; // @todo - fetch properly
+
+    const payload = parseSpaceFormData(formData, alwaysAvailable, availability);
+
+    try {
+      const response = await update(spaceId, payload);
+      console.log("Space updated", response);
+    } catch (error) {
+      console.error("Error updating space:", error);
+    }
+  }
 </script>
 
-<form onsubmit={handleCreateSpace}>
+<form onsubmit={handleSubmit}>
   <fieldset>
     <label for="physical">Physical Location</label>
     <input
